Refactor tests and example into appropriate places
