+++ crates/common/tedge_config_macros/tests/examples.rs
@@ -1,0 +1,12 @@
+use camino::Utf8PathBuf;
+use tedge_config_macros::*;
+
+// The macro invocation generates tests of its own for each example value
+define_tedge_config! {
+    device: {
+        #[tedge_config(example = "/test/cert/path")]
+        #[tedge_config(example = "/test/cert/path2")]
+        #[doku(as = "std::path::PathBuf")]
+        root_cert_path: Utf8PathBuf,
+    }
+}
